+++ test/channel_test.dart
@@ -1,14 +1,15 @@
 import 'package:realtime_client/realtime_client.dart';
+import 'package:realtime_client/src/constants.dart';
 import 'package:test/test.dart';
 
 void main() {
   late RealtimeClient socket;
-  late RealtimeSubscription channel;
+  late RealtimeChannel channel;
 
   const defaultRef = '1';
 
   test('channel should be initially closed', () {
-    final channel = RealtimeSubscription('topic', RealtimeClient('endpoint'));
+    final channel = RealtimeChannel('topic', RealtimeClient('endpoint'));
     expect(channel.isClosed, isTrue);
     channel.rejoin(const Duration(seconds: 5));
     expect(channel.isJoining, isTrue);
@@ -17,13 +18,19 @@ void main() {
   group('constructor', () {
     setUp(() {
       socket = RealtimeClient('', timeout: const Duration(milliseconds: 1234));
-      channel = RealtimeSubscription('topic', socket, params: {'one': 'two'});
+      channel =
+          RealtimeChannel('topic', socket, params: RealtimeChannelConfig());
     });
 
     test('sets defaults', () {
       expect(channel.isClosed, true);
       expect(channel.topic, 'topic');
-      expect(channel.params, {'one': 'two'});
+      expect(channel.params, {
+        'config': {
+          'broadcast': {'ack': false, 'self': false},
+          'presence': {'key': ''}
+        }
+      });
       expect(channel.socket, socket);
     });
   });
@@ -31,13 +38,21 @@ void main() {
   group('join', () {
     setUp(() {
       socket = RealtimeClient('wss://example.com/socket');
-      channel = socket.channel('topic', chanParams: {'one': 'two'});
+      channel = socket.channel('topic');
     });
 
     test('sets state to joining', () {
       channel.subscribe();
 
       expect(channel.isJoining, true);
+    });
+
+    test('sets joinedOnce to true', () {
+      expect(channel.joinedOnce, isFalse);
+
+      channel.subscribe();
+
+      expect(channel.joinedOnce, isTrue);
     });
 
     test('throws if attempting to join multiple times', () {
@@ -48,41 +63,45 @@ void main() {
 
     test('can set timeout on joinPush', () {
       const newTimeout = Duration(milliseconds: 2000);
-      final joinPush = channel.subscribe(timeout: newTimeout);
+      final joinPush = channel.joinPush;
 
-      expect(joinPush.timeout, const Duration(milliseconds: 2000));
+      expect(joinPush.timeout, Constants.defaultTimeout);
+
+      channel.subscribe((_, [__]) {}, newTimeout);
+
+      expect(joinPush.timeout, newTimeout);
     });
   });
 
   group('onError', () {
     setUp(() {
       socket = RealtimeClient('/socket');
-      channel = socket.channel('topic', chanParams: {'one': 'two'});
+      channel = socket.channel('topic');
       channel.subscribe();
     });
 
     test("sets state to 'errored'", () {
-      expect(channel.isErrored, false);
+      expect(channel.isErrored, isFalse);
 
       channel.trigger('phx_error');
 
-      expect(channel.isErrored, true);
+      expect(channel.isErrored, isTrue);
     });
   });
 
   group('onClose', () {
     setUp(() {
       socket = RealtimeClient('/socket');
-      channel = socket.channel('topic', chanParams: {'one': 'two'});
+      channel = socket.channel('topic');
       channel.subscribe();
     });
 
     test("sets state to 'closed'", () {
-      expect(channel.isClosed, false);
+      expect(channel.isClosed, isFalse);
 
       channel.trigger('phx_close');
 
-      expect(channel.isClosed, true);
+      expect(channel.isClosed, isTrue);
     });
   });
 
@@ -90,7 +109,7 @@ void main() {
     setUp(() {
       socket = RealtimeClient('/socket');
 
-      channel = socket.channel('topic', chanParams: {'one': 'two'});
+      channel = socket.channel('topic');
     });
 
     test('returns payload by default', () {
@@ -101,50 +120,48 @@ void main() {
   });
 
   group('on', () {
-    late RealtimeSubscription channel;
+    late RealtimeChannel channel;
 
     setUp(() {
-      channel = RealtimeSubscription('topic', RealtimeClient('endpoint'));
+      channel = RealtimeChannel('topic', RealtimeClient('endpoint'));
     });
 
     test('sets up callback for event', () {
       var callbackCalled = 0;
-      channel.on('event', (dynamic payload, {String? ref}) => callbackCalled++);
+      channel.onEvents('event', ChannelFilter(),
+          (dynamic payload, [dynamic ref]) => callbackCalled++);
 
-      channel.trigger('event', payload: {});
+      channel.trigger('event', {});
       expect(callbackCalled, 1);
     });
 
     test('other event callbacks are ignored', () {
       var eventCallbackCalled = 0;
       var otherEventCallbackCalled = 0;
-      channel.on(
+      channel.onEvents(
         'event',
-        (dynamic payload, {String? ref}) => eventCallbackCalled++,
+        ChannelFilter(),
+        (dynamic payload, [dynamic ref]) => eventCallbackCalled++,
       );
-      channel.on(
+      channel.onEvents(
         'otherEvent',
-        (dynamic payload, {String? ref}) => otherEventCallbackCalled++,
+        ChannelFilter(),
+        (dynamic payload, [dynamic ref]) => otherEventCallbackCalled++,
       );
 
-      channel.trigger('event', payload: {});
+      channel.trigger('event', {});
       expect(eventCallbackCalled, 1);
       expect(otherEventCallbackCalled, 0);
     });
 
     test('"*" bind all events', () {
       var callbackCalled = 0;
-      channel.on('*', (dynamic payload, {String? ref}) => callbackCalled++);
+      channel.onEvents('realtime', ChannelFilter(event: '*'),
+          (dynamic payload, [dynamic ref]) => callbackCalled++);
 
-      channel.trigger('INSERT', payload: {});
-      expect(callbackCalled, 0);
-
-      channel.trigger('*', payload: {'type': 'INSERT'});
-      expect(callbackCalled, 0);
-
-      channel.trigger('INSERT', payload: {'type': 'INSERT'});
-      channel.trigger('UPDATE', payload: {'type': 'UPDATE'});
-      channel.trigger('DELETE', payload: {'type': 'DELETE'});
+      channel.trigger('realtime', {'event': 'INSERT'});
+      channel.trigger('realtime', {'event': 'UPDATE'});
+      channel.trigger('realtime', {'event': 'DELETE'});
       expect(callbackCalled, 3);
     });
   });
@@ -153,7 +170,7 @@ void main() {
     setUp(() {
       socket = RealtimeClient('/socket');
 
-      channel = socket.channel('topic', chanParams: {'one': 'two'});
+      channel = socket.channel('topic');
     });
 
     test('removes all callbacks for event', () {
@@ -161,23 +178,17 @@ void main() {
       var callbackEventCalled2 = 0;
       var callbackOtherCalled = 0;
 
-      channel.on(
-        'event',
-        (dynamic payload, {String? ref}) => callBackEventCalled1++,
-      );
-      channel.on(
-        'event',
-        (dynamic payload, {String? ref}) => callbackEventCalled2++,
-      );
-      channel.on(
-        'other',
-        (dynamic payload, {String? ref}) => callbackOtherCalled++,
-      );
+      channel.onEvents('event', ChannelFilter(),
+          (dynamic payload, [dynamic ref]) => callBackEventCalled1++);
+      channel.onEvents('event', ChannelFilter(),
+          (dynamic payload, [dynamic ref]) => callbackEventCalled2++);
+      channel.onEvents('other', ChannelFilter(),
+          (dynamic payload, [dynamic ref]) => callbackOtherCalled++);
 
-      channel.off('event');
+      channel.off('event', {});
 
-      channel.trigger('event', payload: {}, ref: defaultRef);
-      channel.trigger('other', payload: {}, ref: defaultRef);
+      channel.trigger('event', {}, defaultRef);
+      channel.trigger('other', {}, defaultRef);
 
       expect(callBackEventCalled1, 0);
       expect(callbackEventCalled2, 0);
@@ -185,20 +196,21 @@ void main() {
     });
   });
 
-  group('unsubscribe', () {
+  group('leave', () {
     setUp(() {
       socket = RealtimeClient('/socket');
 
-      channel = socket.channel('topic', chanParams: {'one': 'two'});
-      channel.subscribe().trigger('ok', {});
+      channel = socket.channel('topic');
+      channel.subscribe();
+      channel.joinPush.trigger('ok', {});
     });
 
     test("closes channel on 'ok' from server", () {
-      final anotherChannel =
-          socket.channel('another', chanParams: {'three': 'four'});
+      final anotherChannel = socket.channel('another');
       expect(socket.channels.length, 2);
 
-      channel.unsubscribe().trigger('ok', {});
+      channel.unsubscribe();
+      channel.joinPush.trigger('ok', {});
 
       expect(socket.channels.length, 1);
       expect(socket.channels[0].topic, anotherChannel.topic);
@@ -207,16 +219,18 @@ void main() {
     test("sets state to closed on 'ok' event", () {
       expect(channel.isClosed, false);
 
-      channel.unsubscribe().trigger('ok', {});
+      channel.unsubscribe();
+      channel.joinPush.trigger('ok', {});
 
       expect(channel.isClosed, true);
     });
 
     test("able to unsubscribe from * subscription", () {
-      channel.on('*', (payload, {ref}) {});
+      channel.onEvents('*', ChannelFilter(), (payload, [ref]) {});
       expect(socket.channels.length, 1);
 
-      channel.unsubscribe().trigger('ok', {});
+      channel.unsubscribe();
+      channel.joinPush.trigger('ok', {});
 
       expect(socket.channels.length, 0);
     });